pyproject.toml
[project]
name = "gmail-insighter"
version = "0.1.0"
description = ""
authors = [
    {name = "NIKHIL KUMAR REDDY",email = "nikhilkumar.m@skil.ai"}
]
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "streamlit (>=1.47.1,<2.0.0)",
    "streamlit-google-auth (>=1.1.8,<2.0.0)",
    "pydantic-settings (>=2.10.1,<3.0.0)",
    "fastapi-mcp (>=0.4.0,<0.5.0)",
    "openai (>=1.98.0,<2.0.0)",
    "fastmcp (>=2.10.6,<3.0.0)",
    "autogen-agentchat (>=0.7.1,<0.8.0)",
    "autogen-ext[mcp] (>=0.7.1,<0.8.0)",
    "tiktoken (>=0.9.0,<0.10.0)",
    "langchain-google-community[gmail] (>=2.0.7,<3.0.0)",
    "langchain-community (>=0.3.27,<0.4.0)",
    "loguru (>=0.7.3,<0.8.0)",
    "redis (>=6.2.0,<7.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
